refactor(logger): give each component library its own logger
The logger was a single process-wide instance hosted in one compiled
translation unit, so every solver library shared it. Splitting libcuopt
into components means routing and mathopt should log independently, and
nothing should have to exist purely to host the state.

The logger is now header-only and, crucially, hidden. Hidden visibility
is what does the separating: the static local of an inline function is
emitted as an STB_GNU_UNIQUE symbol, which glibc merges across the whole
process regardless of RTLD_LOCAL, so a header-only logger with default
visibility would still have been one shared instance.

Callers outside the libraries cannot reach a hidden logger, so each
component exports a configure entry point. `init_logger_t` keeps its
meaning -- configure the logger of whichever image constructs it, which
is what the pdlp, mip and grpc solve paths already want -- and the new
`init_component_logger_t` reaches a chosen library from outside. It
defaults to mathopt, so all eight existing external call sites keep
working unchanged, and routing is opted into explicitly.

Two things had to change to make one log file survive several loggers:

- The exported entry point now takes the same ref-count guard that
  `init_logger_t` takes. Without it the MIP solve path reconfigured the
  logger mid-run and, with truncate set, cleared a file the caller had
  already written to.
- File sinks always open in append mode, with a single explicit truncate
  up front. A non-appending sink writes from offset 0 and silently
  overwrites what another logger has appended.

routing::solve now initialises its own logger from the settings. Routing
never constructed one, so its CUOPT_LOG_ERROR calls went into a buffer
that nothing drained and were lost.

Verified: libcuopt.so exports the four entry points and none of the
logger state; cuopt_cli writes both its own and the solver's messages to
one file and still truncates between runs. ctest failures are identical
to clean main in this environment (10 suites, 908 gtest failures, both).
