+++ cpp/src/math_optimization/CMakeLists.txt
@@ -9,6 +9,7 @@ list(PREPEND
   ${CMAKE_CURRENT_SOURCE_DIR}/solution_reader.cu
   ${CMAKE_CURRENT_SOURCE_DIR}/solution_writer.cu
   ${CMAKE_CURRENT_SOURCE_DIR}/tic_toc.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/logger_entry.cpp
   )
 
 set(CUOPT_SRC_FILES ${CUOPT_SRC_FILES}
