+++ cpp/src/routing/CMakeLists.txt
@@ -1,9 +1,10 @@
 # cmake-format: off
-# SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2024-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 # cmake-format: on
 
 set(ROUTING_SRC_FILES
+  ${CMAKE_CURRENT_SOURCE_DIR}/logger_entry.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/local_search/compute_insertions.cu
   ${CMAKE_CURRENT_SOURCE_DIR}/ges/squeeze.cu
   ${CMAKE_CURRENT_SOURCE_DIR}/local_search/sliding_window.cu
