+++ cpp/tests/dual_simplex/unit_tests/solve.cpp
@@ -23,7 +23,7 @@ namespace cuopt::mathematical_optimization::simplex::test {
 
 TEST(dual_simplex, chess_set)
 {
-  cuopt::init_logger_t log("", true);
+  cuopt::init_component_logger_t log("", true);
   namespace simplex = cuopt::mathematical_optimization::simplex;
   raft::handle_t handle{};
   simplex::user_problem_t<int, double> user_problem(&handle);
@@ -97,7 +97,7 @@ TEST(dual_simplex, chess_set)
 
 TEST(dual_simplex, burglar)
 {
-  cuopt::init_logger_t log("", true);
+  cuopt::init_component_logger_t log("", true);
   constexpr int num_items     = 8;
   constexpr double max_weight = 102;
 
@@ -173,7 +173,7 @@ TEST(dual_simplex, burglar)
 
 TEST(dual_simplex, empty_columns)
 {
-  cuopt::init_logger_t log("", true);
+  cuopt::init_component_logger_t log("", true);
   // Same as burglar problem above but with an empty column inserted
   constexpr int num_items     = 9;
   constexpr double max_weight = 102;
@@ -262,7 +262,7 @@ TEST(dual_simplex, empty_columns)
 
 TEST(dual_simplex, dual_variable_greater_than)
 {
-  cuopt::init_logger_t log("", true);
+  cuopt::init_component_logger_t log("", true);
   // minimize   3*x0 + 2 * x1
   // subject to  x0 + x1  >= 1
   //             x0 + 2x1 >= 3
